+++ HISTORY.md
@@ -26,10 +26,10 @@
 
 0.2.5 / 27-09-2020
 ====================
-	* Implemented metaData.getKuzs()
-	* Implemented metaData.getCodeFiles()
+	* Implemented `metaData.getKuzs()`
+	* Implemented `metaData.getCodeFiles()`
 	* Now not watching by default
-	* Renamed kuz-fs to kuz-fsutils
+	* Renamed `kuz-fs` to `kuz-fsutils`
 
 
 0.2.4 / 27-09-2020