+++ HISTORY.md
@@ -24,6 +24,7 @@
 	* Now passing `k` and `s` objects
 	* Now passing `layout` and `theme`
 	* Added new keywords
+	* Added `KuzCodeFile.getContent()`
 
 
 0.2.5 / 27-09-2020